refactor : Amount 위치 변경
