+++ src/main/kotlin/view/Amount.kt
@@ -1,4 +1,4 @@
-package model.money
+package view
 
 internal const val EMPTY_VALUE_ERROR_MESSAGE = "값이 비어있습니다."
 internal const val INVALID_NUMBER_FORMAT_MESSAGE = "입력한 값은 숫자여야 합니다."
